fix imports / types
