+++ src/types/k8s.go
@@ -1,0 +1,18 @@
+package types
+
+// ZarfState is maintained as a secret in the Zarf namespace to track Zarf init data
+type ZarfState struct {
+	ZarfAppliance bool         `json:"zarfAppliance"`
+	Distro        string       `json:"distro"`
+	Architecture  string       `json:"architecture"`
+	StorageClass  string       `json:"storageClass"`
+	Secret        string       `json:"secret"`
+	NodePort      string       `json:"nodePort"`
+	AgentTLS      GeneratedPKI `json:"agentTLS"`
+}
+
+type GeneratedPKI struct {
+	CA   []byte
+	Cert []byte
+	Key  []byte
+}
